requirements.txt
# Ultimate SMC Trading Bot - Dependencies
# Python 3.10+ required

# MetaTrader 5 Connection
MetaTrader5>=5.0.45

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0
python-decouple>=3.8

# Visualization
plotly>=5.18.0
mplfinance>=0.12.10b0

# Technical Analysis
ta>=0.11.0
pandas-ta>=0.3.14b

# Utilities
colorama>=0.4.6
tabulate>=0.9.0
rich>=13.7.0

# Logging
loguru>=0.7.2

# Scheduling
schedule>=1.2.1

# Date/Time
python-dateutil>=2.8.2
pytz>=2024.1

# HTTP Requests (for news filter)
requests>=2.31.0

# Web Dashboard (optional)
flask>=3.0.0

# Statistics (optional)
scipy>=1.12.0

# Discord Notifications (optional)
python-dotenv>=1.0.0

# Optional: Backtesting
# backtesting>=0.3.3

# --- Fundamental Analysis (NOUVEAU) ---
# Economic Calendar & News
investpy>=1.0.8

# Market Data (DXY, VIX, Yields)
yfinance>=0.2.32

# Testing (Phase 1)
pytest>=7.4.3
pytest-cov>=4.1.0
